% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enaho-package.R
\docType{package}
\name{enaho-package}
\alias{enaho}
\alias{enaho-package}
\title{enaho: Build A Package From Rmarkdown File}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\author{
\strong{Maintainer}: Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
