% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulos.desc.R
\name{modulos.desc}
\alias{modulos.desc}
\title{Módulos de una encuesta}
\usage{
modulos.desc(encuesta = "ENAHO", modulos = NULL, mostrarannos = FALSE)
}
\arguments{
\item{encuesta}{el nombre de una encuesta del INEI. Las opciones dispobibles son
\code{"ENAHO"} y \code{"ENAHOpanel"}. No importa si está escrito en mayúscula o minúscula.}

\item{modulos}{el código de los módulos que mostrar. Si es \code{NULL} muestra todos
los disponibles.}

\item{mostrarannos}{un valor lógico que indica si se deben mostrar los años disponibles.}
}
\value{
un data frame.
}
\description{
Muestra los módulos de una encuesta y en que año se encuentran disponibles.
}
\examples{
modulos.desc()

}
