% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulos.R
\name{modulos}
\alias{modulos}
\title{Módulos de una encuesta}
\usage{
modulos(encuesta = "ENAHO", modulo = NULL)
}
\arguments{
\item{encuesta}{el nombre de una encuesta del INEI. Las opciones dispobibles son
\code{"ENAHO"} y \code{"ENAHOpanel"}. No importa si está escrito en mayúscula o minúscula.}

\item{modulo}{el código de los módulos que mostrar. Si es \code{NULL} muestra todos
los disponibles. Si no es \code{NULL} muestra los períodos disponibles.}
}
\value{
un data frame.
}
\description{
Muestra los módulos de una encuesta y en que año se encuentran disponibles.
}
\examples{
modulos()

}
