% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporcion.enaho.R
\name{proporcion.enaho}
\alias{proporcion.enaho}
\title{Proporción}
\usage{
proporcion.enaho(
  x,
  base,
  estratos = NULL,
  combinarestratos = FALSE,
  dividirperiodos = FALSE,
  ruralidad = 6:8,
  pesos = "FACTOR07",
  forzaretiquetas = FALSE,
  porcentaje = FALSE,
  decimales = NULL,
  formatolargo = FALSE
)
}
\arguments{
\item{x}{una cadena de caracteres con los nombres de las variables a estimar.}

\item{base}{una base leída con \code{\link{leer.inei}} o \code{\link{leer.inei.web}} en forma de tibble
o en forma de lista.}

\item{estratos}{una cadena de caracteres con los nombres de las variables que deben ser consideradas como estratos.
Además se puede usar \code{"area"} y \code{"region"}, con las que se calcularán los resultados para los
estratos urbano y rural, y para cada región.}

\item{combinarestratos}{un valor lógico que indica si debe calcularse también la combinación entre estratos.
Por ejemplo, si los estatos son \code{"area"} y \code{"region"}, se estimará también para cada área de
cada región.}

\item{dividirperiodos}{un valor lógico que indica si los resultados deben ser separados por período.}

\item{ruralidad}{un vector numérico que indica que valores de la variable \code{"ESTRATO"} deben ser considerados
rurales. Usualmente del 6 al 8.}

\item{pesos}{una cadena de caracteres con el nombre de la variable que representa los pesos.
Si es \code{NULL} las proporciones se estiman sin pesarlas.}

\item{forzaretiquetas}{un valor lógico que indica si las etiquetas del período con más etiquetas deben
ser forzadas para todos los períodos. No es poco común que las etiquetas de una variable cambien con los
años, si esto ocurre, esta función las interpretará como distintas. Por ejemplo \code{"INCIAL"} y
\code{"EDUCACION INICIAL"}. Sin embargo, usando este argumento se puede forzar a que se interpreten los
mismos valores todos los años. Usar con discreción, y, si es posible, siempre después de una exploración
de datos usando \code{forzaretiquetas = FALSE}.}

\item{porcentaje}{un valor lógico que indica si los resultados deben presentarse como porcentaje en vez de
proporción.}

\item{decimales}{un valor numérico que indica cuántos decimales se necesitan. Si es \code{NULL} no se redondea.}

\item{formatolargo}{un valor lógico que indica si los resultados debes estar en formato largo.
Si se incluye más de una variable, se darán los resultados en formato largo.}
}
\value{
un data frame o una lista.
}
\description{
Estima proporciones de variables en bases leídas con \code{\link{leer.inei}} o
\code{\link{leer.inei.web}}.
}
\examples{
directorio = system.file("extdata", package = "enaho")
leido <- leer.inei(modulo = 37, periodos = 2011, directorio = directorio, tipo = c("t1","t2"))
proporcion.enaho(x = c("P7061"),base = leido,pesos = "FACTRIM")

}
